Fix how the wallet network is handled
Now that the wallet network is stored in `keys.txt`, it is invalid to
pass around network parameters alongside the wallet directory. They
are now parsed at the place where the wallet directory is used (often at
the same time as parsing the rest of `keys.txt`), and then passed around
with other variables derived from it.

This also fixes a bug with the `init` command, where a new wallet could
not be created unless an existing wallet at that path existed (to parse
the network parameters from), but if that path existed then `init`
(correctly) refused to overwrite it. The network parameters for this
command are now correctly constructed from the config option.