sync: Refresh UTXOs at the start of each scanning cycle
The new `transparent-inputs` feature flag is default-enabled because
most testing of the Rust crates that we want to do will be with that
flag enabled, and anyone using this tool to investigate existing wallet
databases will also need it enabled. But the flag can be disabled for
testing shielded-only workflows.

Closes Electric-Coin-Company/zec-sqlite-cli#21.